refactor: 예외처리 분리 (#35)
